requirements.txt
discord.py==1.6.0
inflect==5.2.0
twitchio==1.2.1
